fix(arch): move AssistantConfigTypes to LangTeach.Api.AI; add ValidateProposalFields
ProposalFieldsConfig and entry types matched the PromptFragmentsConfig pattern
(config shapes loaded at startup) so they belong in the AI namespace, not DTOs.
Add ValidateProposalFields startup guard so a corrupted proposal-fields.json
fails fast on startup rather than silently emitting zero proposals.
