+++ backend/LangTeach.Api/Services/IPedagogyConfigService.cs
@@ -1,6 +1,5 @@
 using System.Collections.Frozen;
 using LangTeach.Api.AI;
-using LangTeach.Api.DTOs;
 
 namespace LangTeach.Api.Services;
